feat: add http get support
